renaming specialized implementations appropriately
